% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso3-data.R
\docType{data}
\name{iso3}
\alias{iso3}
\title{List of available countries with their ISO 3166-1 alpha-3 codes}
\format{An object of class \code{data.frame} with 254 rows and 3 columns.}
\description{
The list of countries with available maps at \url{http://gadm.org/}
}
\usage{
data(iso3)
}
\examples{
data(iso3)
str(iso3)
}
\references{
\url{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}
}
\keyword{datasets}
\keyword{alpha3}
\keyword{map}
